// packages/contracts/generated/AnnotationDocument.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnnotationObject } from "./AnnotationObject";
import type { Completion } from "./Completion";
import type { CoordinateSpace } from "./CoordinateSpace";
import type { Id } from "./Id";

export type AnnotationDocument = { schema_version: 1, asset_revision_id: Id, ontology_version_id: Id, coordinate_space: CoordinateSpace, completion: Completion, objects: Array<AnnotationObject>, };